requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.6.0
google-genai>=0.3.0
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0
Pillow>=10.0.0
piexif>=1.1.3
python-multipart>=0.0.9
